Move page content to model.py
